requirements.txt
apache-airflow==1.9.0
cryptography==2.3.1
flask-bcrypt==0.7.1
sqlalchemy<1.2